backend/requirements.txt
fastapi
uvicorn
pandas
scikit-learn
python-multipart
python-dotenv
requests
fpdf
joblib
matplotlib
seaborn
